Persist pitch_detector instance across blocks
I was creating a new pitch_detector for each audio block prior to this
which wasn't ideal. This commit moves the pitch_detector to the
initialise method. I needed to use a `shared_ptr` to work around the
need to initialise it and then set the params again afterwards due to
how vamp works.

Tested with Sonic Visualiser and looking good now.