Add regularOutput mode and compensate for offset in pitch detection
The algorithm require 2*lowest_frequency.period() before it can output a
detection. This introduces a lag (common to all autocorrelation methods)
- the vamp plugin compensates this by shifting the timestamps back by a
  constant amound (2*lowest_frequency.period()).

This also introduces a regularOutput mode where a pitch/timestamp pair
are output for every n frames. The defaults is 128 to
match the output of the MedleyDB dataset for easier comparison.